First implementation of driver with Hittable
