+++ src/main.rs
@@ -1,22 +1,17 @@
 mod ray_tracing;
-use ray_tracing::{color::*, ray::Ray};
 
-use crate::ray_tracing::{point::Point, vec3::Vec3};
+use ray_tracing::{color::*, hittable::Hittable, ray::Ray};
+
+use crate::ray_tracing::{point::Point, sphere::Sphere, vec3::Vec3};
 
 fn header(width: i32, height: i32) -> String {
     format!("P3\n{width} {height}\n255")
 }
 
-fn ray_color(ray: Ray) -> Color {
-    let center = Point::new(0.0, 0.0, -1.0);
-    match intersection_points(center, 0.5, ray) {
-        // Take the first point. In this case we know if there is an intersection, first solution is closer
-        (Some(t), _) => {
-            let intersect_point = ray.at(t);
-            let intersect_normal = Vec3::from(intersect_point - center).unit();
-            Color::from(intersect_normal + Vec3::new(1.0, 1.0, 1.0)) * 0.5
-        }
-        _ => {
+fn ray_color<T: Hittable + ?Sized>(ray: Ray, world: &T) -> Color {
+    match world.hit(&ray, 0.0..) {
+        Some(hit_data) => 0.5 * (Color::from(hit_data.normal) + Color::new(1.0, 1.0, 1.0)),
+        None => {
             let unit_dir = ray.dir.unit();
             let t = 0.5 * (unit_dir.y + 1.0);
             (1.0 - t) * Color::new(1.0, 1.0, 1.0) + t * Color::new(0.5, 0.7, 1.0)
@@ -24,40 +19,26 @@ fn ray_color(ray: Ray) -> Color {
     }
 }
 
-type Solutions = (Option<f32>, Option<f32>);
-fn intersection_points(center: Point, radius: f32, ray: Ray) -> Solutions {
-    // Equation of points on ray that are on sphere centered at C with radius r is:
-    // 𝑡^2*𝐛⋅𝐛 + 2𝑡*𝐛⋅(𝐀−𝐂) + (𝐀−𝐂)⋅(𝐀−𝐂) − 𝑟^2 = 0
-    // => discriminant = b^2 - 4*a*c
-    // if discriminant > 0 => there are solutions to the equation (possibly negative)
-    let shifted_center = Vec3::from(ray.origin - center);
-    let a = ray.dir.dot(ray.dir);
-    let half_b = ray.dir.dot(shifted_center);
-    let c: f32 = shifted_center.dot(shifted_center) - radius * radius;
-    let discriminant = half_b * half_b - a * c;
-    if discriminant < 0.0 {
-        return (None, None);
-    }
-    // This is the 'smaller' solution, in terms of t
-    let solution_a = (-half_b - f32::sqrt(discriminant)) / a;
-    let solution_b = (-half_b + f32::sqrt(discriminant)) / a;
-    (
-        if solution_a >= 0.0 {
-            Some(solution_a)
-        } else {
-            None
-        },
-        if solution_b >= 0.0 {
-            Some(solution_b)
-        } else {
-            None
-        },
-    )
-}
-
 fn main() {
+    // Image
     let aspect_ratio: f32 = 16.0 / 9.0;
     let (image_width, image_height) = (400, (400.0 / aspect_ratio).trunc() as i32);
+
+    // World
+
+    let world = vec![
+        Sphere {
+            radius: 0.5,
+            center: Point::new(0.0, 0.0, -1.0),
+        },
+        Sphere {
+            radius: 100.0,
+            center: Point::new(0.0, -100.5, -1.0),
+        },
+    ];
+
+    // Camera
+
     let viewport_height = 2.0;
     let viewport_width = aspect_ratio * viewport_height;
     let focal_len = 1.0;
@@ -68,6 +49,8 @@ fn main() {
         - (horizontal / 2.0).into()
         - (vertical / 2.0).into()
         - Vec3::new(0.0, 0.0, focal_len).into();
+
+    // Render
 
     let header = header(image_width, image_height);
     println!("{header}");
@@ -81,7 +64,7 @@ fn main() {
                 origin,
                 Vec3::from(lower_left) + u * horizontal + v * vertical - Vec3::from(origin),
             );
-            let color = ray_color(ray);
+            let color = ray_color(ray, world.as_slice());
             println!("{color}");
         }
     }
