+++ src/ray_tracing/mod.rs
@@ -2,5 +2,6 @@ pub mod color;
 pub mod hittable;
 pub mod point;
 pub mod ray;
+pub mod sphere;
 mod test_utils;
 pub mod vec3;
